expose auth state via callback
