+++ packages/core/src/redux/features/session/sessionSaga.ts
@@ -13,6 +13,7 @@ import type { SessionChannel, SwEventChannel } from '../../interfaces'
 import { createCatchableSaga } from '../../utils/sagaHelpers'
 import { socketMessageAction } from '../../actions'
 import { getLogger, isWebrtcEventType, toInternalAction } from '../../../utils'
+import { sessionActions } from './sessionSlice'
 
 type SessionSagaParams = {
   session: BaseSession
@@ -54,10 +55,15 @@ export function* sessionChannelWatcher({
     /**
      * After connecting to the SignalWire network, it sends the `authorization_state`
      * through the `signalwire.authorization.state` event. We store this value in
-     * the storage (or expose it to the client) since it is required for reconnect.
+     * the browser storage for JWT and in Redux store for SAT since it is required for reconnect.
      */
     if (isSwAuthorizationState(broadcastParams)) {
       session.onSwAuthorizationState(broadcastParams.params.authorization_state)
+      yield put(
+        sessionActions.updateAuthStateEncoded(
+          broadcastParams.params.authorization_state
+        )
+      )
       return
     }
 
@@ -65,6 +71,8 @@ export function* sessionChannelWatcher({
      * Put actions with `event_type` to trigger all the children sagas
      * This should replace all the isWebrtcEvent/isVideoEvent guards below
      * since we'll move that logic on a separate package.
+     *
+     * TODO: We no longer need this and it can be removed
      */
     yield put({ type: broadcastParams.event_type, payload: broadcastParams })
   }
